requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.18
sentence-transformers==2.2.2
pydantic==2.5.2
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
loguru==0.7.2
pandas==2.1.4
numpy==1.24.3
openai==1.6.1
